#221 refactor: 투두 관련 스타일 분리
