+++ src/components/Todo/TodoItem.tsx
@@ -1,15 +1,28 @@
 import { Check, Close, Delete, Edit } from '@/assets/icons';
 import { Button } from '@/components/common';
 import IconWrapper from '@/components/IconWrapper';
+import {
+  actionStyle,
+  cancelButtonStyle,
+  checkedStyle,
+  checkIconStyle,
+  completedContentStyle,
+  contentWrapperStyle,
+  deleteButtonStyle,
+  deleteModeStyle,
+  editActionStyle,
+  editInputStyle,
+  getStatusStyles,
+  todoItemStyle,
+} from '@/components/Todo/styles';
 import { TODO_STATUS } from '@/constants/todo';
 import { useToastHandler } from '@/hooks';
 import { useDeleteTodo, useUpdateTodo } from '@/hooks/mutations';
 import { todoKeys } from '@/hooks/queries';
 import { useDayStore, useSquadStore } from '@/stores';
-import { pcMediaQuery } from '@/styles/breakpoints';
 import { fullSizeButtonStyle } from '@/styles/globalStyles';
 import { ToDoDetail, TodoQueryParams, UpdateTodoRequest } from '@/types';
-import { css, keyframes, Theme, useTheme } from '@emotion/react';
+import { useTheme } from '@emotion/react';
 import { useQueryClient } from '@tanstack/react-query';
 import { KeyboardEvent, useState } from 'react';
 
@@ -173,152 +186,3 @@ const TodoItem = ({ todo, squadMemberId }: { todo: ToDoDetail; squadMemberId: nu
 };
 
 export default TodoItem;
-
-export const todoItemStyle = css`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 8px;
-  border-radius: 16px;
-  margin: 0 8px;
-  padding: 8px 0 8px 8px;
-  min-height: 48px;
-  flex-shrink: 0;
-  cursor: pointer;
-  transition: transform 0.3s ease-out;
-
-  & button,
-  div {
-    font-size: 0.8rem;
-  }
-
-  ${pcMediaQuery(css`
-    & button,
-    div {
-      font-size: 1rem;
-    }
-  `)}
-`;
-
-export const contentWrapperStyle = css`
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  height: 100%;
-`;
-
-export const checkIconStyle = (theme: Theme) => css`
-  min-width: 20px;
-  max-width: 20px;
-  height: 20px;
-  background-color: ${theme.colors.background.white};
-  border: 2px solid ${theme.colors.gray.gray100};
-  border-radius: 8px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const checkedStyle = (theme: Theme) => css`
-  background-color: ${theme.colors.primary};
-  border: none;
-
-  & svg {
-    color: white;
-    stroke-width: 2;
-    width: 12px;
-    height: 12px;
-  }
-`;
-
-export const getStatusStyles = (isChecked: boolean, theme: Theme) => {
-  switch (isChecked) {
-    case true:
-      return css`
-        outline: 1.5px solid ${theme.colors.primary};
-        background-color: ${theme.colors.background.lightYellow};
-      `;
-    default:
-      return css`
-        background-color: ${theme.colors.background.white};
-      `;
-  }
-};
-
-const completedContentStyle = (theme: Theme) => css`
-  color: ${theme.colors.gray.gray200};
-  text-decoration: line-through ${theme.colors.gray.gray200};
-  text-decoration-thickness: 1.5px;
-`;
-
-const slideInFromRight = keyframes`
-  from {
-    transform: translateX(100%);
-  }
-  to {
-    transform: translateX(0);
-  }
-`;
-
-const deleteModeStyle = css`
-  animation: ${slideInFromRight} 0.3s ease-out;
-
-  & button {
-    font-weight: 500;
-    text-align: center;
-    border-radius: 12px;
-    color: white;
-  }
-`;
-
-const deleteButtonStyle = css`
-  background-color: #ff5a5a;
-`;
-
-const cancelButtonStyle = css`
-  width: 56px;
-  height: 100%;
-  background-color: #dfdfdf;
-  border-radius: 8px;
-  margin-right: 8px;
-
-  & svg {
-    width: 20px;
-    color: var(--color-text-gray);
-  }
-`;
-
-const actionStyle = css`
-  display: flex;
-  gap: 4px;
-  margin-right: 8px;
-
-  & button {
-    color: #bbbbbb;
-  }
-`;
-
-const editInputStyle = (theme: Theme) => css`
-  height: 100%;
-  flex: 1;
-  font-size: 0.8rem;
-  color: var(--color-text-gray);
-
-  ${pcMediaQuery(css`
-    ${theme.typography.size_16}
-  `)}
-`;
-
-const editActionStyle = css`
-  width: 132px;
-  display: flex;
-  gap: 4px;
-  margin-right: 8px;
-  text-decoration: none;
-
-  & button {
-    border-radius: 12px;
-    text-align: center;
-    font-weight: 500;
-  }
-`;
