+++ src/components/Todo/TodoList.tsx
@@ -1,8 +1,8 @@
 import { MemberTodoItem } from '@/components/Member';
+import { memberTodoListStyle, todoContainerStyle } from '@/components/Todo/styles';
 import TodoItem from '@/components/Todo/TodoItem';
 import { useInfinite } from '@/hooks';
 import { ToDoDetail } from '@/types';
-import { css, Theme } from '@emotion/react';
 
 type Props = {
   todos: ToDoDetail[];
@@ -33,21 +33,3 @@ export const TodoList = ({ todos, loadMoreTodos, hasNextPage, isMeSelected, squa
     </ul>
   );
 };
-
-export const todoContainerStyle = (theme: Theme) => css`
-  background-color: ${theme.colors.background.gray};
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  flex-grow: 1;
-  margin: 0 16px;
-  padding: 8px 0;
-  border-radius: 24px;
-  overflow-y: auto;
-  -webkit-box-shadow: 0px 0px 6px 0px rgba(0, 0, 0, 0.1);
-  box-shadow: 0px 0px 6px 0px rgba(0, 0, 0, 0.1);
-`;
-
-const memberTodoListStyle = css`
-  margin-bottom: 24px;
-`;
